Python: Set Workspace Interpreter command
This adds a new Command Palette entry for Python which provides
suggestions for Python interpreters you may wish to set in the workspace
settings. For now, only conda environments are detected (by
interrogating the conda binary), but it will be easy to add support for
other common Python interpreter locations.

VSCode doesn't yet include support for setting workspace settings from
an extension, so for now just a message is printed to demonstrate where
we would update settings.json.