Create localization of strings (#2757)
Creation of a new module - localize. 
This doesn't use the vscode-nls and vscode-nls-dev as they
interfere with our code coverage.
Adding new strings requires adding them to the localize.ts and the
localize.nls.<locale>.json files.
Future work would be to translate the nls.json files into xlf files for
the loc team to own. We'd then write a build time task to turn the
xlf files into the output nls.<locale>.json files so that this code
works the same.

For #463