Enable on-type formatting from language server (#2737)
For #2690.

Similar to #2612. Note that `:`-based formatting isn't implemented in language server yet (Microsoft/python-language-server#165), but it hasn't been enabled in the extension for anyone for a while now either (fixed in #2714/#2724).

Since line formatting is relatively new in language server, this likely shouldn't make it out into a release until the language server downloaded by default is new enough to include that functionality. I haven't been monitoring that side of things, so feel free to hold off on merging this until the right time.