Add "Run All/Debug All" dropdown picker for parametrized tests (#8757)
* Register new command for running parametrized tests

* Show "Run All" or "Debug All" in dropdown picker on parametrized tests

* Use CommandSource as input to second call of getNamesAndValues

In the picker unit tests two calls to getNamesAndValues are used to
extract values from enums. The second call used a generic with type
'CommandSource' but as input a Type object. To be more meaningful this
has been changed to 'CommandSource' so that the test checks different
TestItem types against different command sources. Otherwise it would
test TestItem types against TestItem types.

Furthermore extracting name and value of CommandSource.commandpalette
with getNamesAndValues was not successful and returned undefined. This
has been fixed by assigning CommandSource.commandpalette directly to
commandSource.value when this command source is evaluated.

* Add test for Tests_Run_Parametrized command

* Add news entry for #5608