Drop PythonEnvInfo.defaultDisplayName. (#15241)
The property became irrelevant with #15130, so we should have dropped it there. We're dropping it here after the fact instead.

Note that this change also corrects a bug in the Windows registry locator that I found while dropping PythonEnvInfo.defaultDisplayName. The "display name" found in the registry was getting set to PythonEnvInfo.defaultDisplayName when it should have been set to PythonEnvInfo.distro.defaultDisplayName.