Remove debugger banner (#19849)
### Problem
#### Issue Link
https://github.com/microsoft/vscode-python/issues/19714 
#### Summary
Remove debugger banner
#### Description
It was reported in the issue 19714 that we display a prompt to get
feedback on a separate survey for the debugger, but given we get a lot
of feedback for the debugger in our "general" survey, it should be fine
to remove it.

### Solution
#### Implementation
The current implementation of the debugger survey is defined in a banner
class. This class pops up a banner to ask if the user wants to fill out
the survey or not. If selected yes, a browser with the survey link is
opened.
After discussing with the team, we decided to remove the banner
completely, and if it is needed in the future the team will restore it.
#### Testing
##### Unit test

[debugging-output.txt](https://github.com/microsoft/vscode-python/files/9586036/debugging-output.txt)