Ensure that search location is stored as a string (#20106)
For https://github.com/microsoft/vscode-python/issues/20104

The Cause: The `workspaceFolder` field is set using an internal field
`searchLocation`. `searchLocation` is what gets saved to Memento as
`Uri`. when we read it back it looks like this:

![image](https://user-images.githubusercontent.com/3840081/198420406-fb4c9b87-1276-4717-8043-69d95dcdba48.png)

Interim Solution:
* Reading (broken case): Since there are already existing cases where
this is broken. The surgical fix is to read bad value, extract the
`scheme` and `path`, use that to re-create the URI.
* Reading/Writing (new case): Make sure that we convert Uri to string
before writing and parse back the Uri when reading.

Long term solution:
Object written to memento needs to be serialized using a custom
serializer that validates all values before writing and after reading.