Accept pytest exit code 1 for some tests cannot be collected (#20074)
In case you have some tests where not successfully collected (e.g. 10 ok
and 1 fail) vscode display only an error in the test explorer. For
pytest it is possible to set the flag `--continue-on-collection-errors`
but also vscode only display an error and not the other successfully
collected tests.
Pytest has a specific exit code for this behaviour (1)
[https://docs.pytest.org/en/7.1.x/reference/exit-codes.html](https://docs.pytest.org/en/7.1.x/reference/exit-codes.html)

Co-authored-by: Eleanor Boyd <eleanorboyd@microsoft.com>