Share telemetry from core extension. (#20893)
it turns out the new telemetry API removed a way to set extension id and
version when telemetry reporter is created and it implicitly sets from
extension reporter is created. the same way how LSP client is working.

since we want to keep using the same extension id and etc for our
telemetry, we need the reporter created from core ext.