fix pytest run script to add cwd to path (#21399)
fixes https://github.com/microsoft/vscode-python/issues/21397. The issue
came from the setup of running a script that then runs pytest which is
different then pytest discovery which runs pytest as a module. The run
script didn't work since the script did not insert the cwd into the path
like what is done when running a module. This change inserts the cwd.