Let pylance use SharedArrayBuffer cancellation in the browser (#21482)
Dirk implemented a new cancellation method for the browser in the LSP
library. About 6 months ago.

We use this in the browser to provide cancellation.

Since we moved creation of the pylance language client to Python, this
is picking up the LSP library currently in the Python extension and it
was before this change by Dirk.

This change moves the LSP library up to the latest.