Feature branch testing overflow bug fix (#21812)
This merges in two PRs that were reverted because of a bug introduced
that caused subprocess overflow.
reverted PRs: https://github.com/microsoft/vscode-python/pull/21667,
https://github.com/microsoft/vscode-python/pull/21682

This now implements these two PRs allowing for absolute testIds and an
execObservable for the subprocess. This PR also adds a bug fix and
functional tests to ensure this doesn't happen again.

Since this PR is large, all items in it have already been reviewed as
they were merged into the feature branch.