Always prepend to PATH instead of replacing it (#21906)
For #20822 #11039

Replacing as-is has its problems, for eg. pyenv asks their users to
manipulate `PATH` in their init script:
https://github.com/pyenv/pyenv#set-up-your-shell-environment-for-pyenv,
which we could end up replacing.


![image](https://github.com/microsoft/vscode-python/assets/13199757/cc904f76-8d42-47e1-a6c8-6cfff6543db8)

Particularly for pyenv, it mean users not being able to find pyenv:


![image](https://github.com/microsoft/vscode-python/assets/13199757/26100328-c227-435b-a4f2-ec168099f4c1)

Prepending solves it for cases where initial PATH value is suffix of the
final value:


![image](https://github.com/microsoft/vscode-python/assets/13199757/a95e4ffd-68dc-4e73-905e-504b3051324f)

But, in other cases, this means that we end up with the whole `PATH`
thrice. This is because it prepends it twice:
- Once in shell integration script
- Once when creating a process

So the final value could be:
```
PATH=<activated_full_path><activated_full_path><original_path>
```
where `<activated_full_path>` refers to value of `PATH` environment
variable post activation. eg.


![image](https://github.com/microsoft/vscode-python/assets/13199757/7e771f62-eb53-49be-b261-d259096008f3)