Handle white spaces for list along with dictionary (#22209)
Legacy normalization script leaves unnecessary white spaces for
dictionary as well as list. While there are multiple correct usage of it
such as for after a function, for more intuitive REPL experience. We
want to keep previous normalization style and white space format EXCEPT
for dictionary and list case.

Dictionary case is handled, but this is the PR to handle the elimination
of extra white spaces for list as well.
Closes: #22208