Add consoleTitle to launch.json properties schema (#22406)
consoleTitle is effectively an undocumented feature. pydebug accepts the
config key, but vscode will report a problem because it is missing from
the schema. see [this debugpy
issue](https://github.com/microsoft/debugpy/issues/1178) and
[#13040](https://github.com/microsoft/vscode-python/issues/13040)

I used the phrase "console or terminal" because I noticed the name was
represented on both integratedTerminal and internalConsole. I didn't
test externalTerminal.

---------

Co-authored-by: Luciana Abud <45497113+luabud@users.noreply.github.com>