Fix for issue with environment path and Jedi (#22713)
Fixes https://github.com/microsoft/vscode-python/issues/22659
Fixes https://github.com/microsoft/vscode-python/issues/22672

Jedi checks if the environment path is a path to a file or directory. If
it is a path, it uses it as is. If not it looks for a binary assuming it
is a virtual environment. Conda environments are structured differently
and it fails to find the binary, if we pass in the path to environment
directory.

The fix here is to always pass the path to the binary.